Updated V1.3
+ Added customcommands
+ Added leaveid
+ Added logs

- Added permission check to spp
- Updated help command
- Updated changelog

+++ cogs/core.py
@@ -1,7 +1,10 @@
 import datetime
 import discord
 import asyncio
+import json
 import time
+import re
+
 
 from utils import default
 from utils.default import lib
@@ -12,6 +15,8 @@ start_time = time.time()
 class Core(commands.Cog):
     def __init__(self, bot):
         self.bot = bot
+        with open("./data/customcommands/commands.json") as f:
+            self.cc = json.load(f)
 
     @commands.group(invoke_without_command=True, no_pm=True)
     @commands.cooldown(1, 5, commands.BucketType.user)
@@ -25,11 +30,11 @@ class Core(commands.Cog):
             )
         embed.set_author(name="Devolution", icon_url="https://i.imgur.com/BS6YRcT.jpg")
         embed.set_footer(text="Devolution | Help", icon_url="https://i.imgur.com/BS6YRcT.jpg")
-        embed.add_field(name="Information", value="**help** - Gives help!\n**bug** - Use it to report bugs.\n**sinfo** - Displays guild information.\n**uinfo** - Displays user information\n**uptime** - Displays the bots uptime\n**about** - Displays stuff about the bot\n**changelog** - Displays the entire bots changelog\n**github** - Provides github link", inline=False)
+        embed.add_field(name="Information", value="**help** - Gives help!\n**help permissions** - Gives a list of permissions the bot requires to function\n**bug** - Use it to report bugs.\n**sinfo** - Displays guild information.\n**uinfo** - Displays user information\n**uptime** - Displays the bots uptime\n**about** - Displays stuff about the bot\n**changelog** - Displays the entire bots changelog\n**github** - Provides github link", inline=False)
         embed.add_field(name="Fun", value="**coinflip** - Flip a coin\n**space** - Get live information about the ISS\n**colour** - Get a random colour\n**roll** - Roles a dice\n**insult** - Insult people you dislike!\n**boobs** - See some melons!\n**ass** - See some peaches!\n**gif** - Search up a gif on giphy by name\n**gifr** - Gives a random gif from giphy\n**owo** - Get random responses", inline=False)
-        embed.add_field(name="Moderation", value="**kick** - Kick a mentioned user\n**ban** - Ban a mentioned user\n**hackban** - Allows you to ban a UserID\n**punish** - Gives mute options\n**cleanup** - Gives message moderation options\n**clean** - Deletes the last 100 command messages and bot messages", inline=False)
-        embed.add_field(name="Useful", value="**say** - Speak as the bot\n**rename** - Change a users nickname\n**invite** - Gives usage details\n**embed** - Creates an embed message\n**role** - Gives role options\n**music** - Gives music help", inline=False)
-        embed.add_field(name="Admin", value="**leave** - Makes the bot leave the guild\n**setpresence(sp)** - Change the playing status of the bot.\n**shutdown** - Sends the bot into a deep sleep ...\n**cog** - Displays list of Cog Options\n**todo** - Displays List of shit todo\n**pm** - PMs Target user as bot\n**pmid** - PMs target ID as bot\n**amiadmin** - Tells you if your UserID is inside the cfg file.", inline=False)
+        embed.add_field(name="Useful", value="**say** - Speak as the bot\n**rename** - Change a users nickname\n**invite** - Gives usage details\n**embed** - Creates an embed message\n**role** - Gives role options\n**music** - Gives music help\n**customcommand** - Add customcommands to your server", inline=False)
+        embed.add_field(name="Moderation", value="**kick** - Kick a mentioned user\n**ban** - Ban a mentioned user\n**hackban** - Allows you to ban a UserID\n**punish** - Gives mute options\n**cleanup** - Gives message moderation options\n**clean** - Deletes the last 100 command messages and bot messages\nlogs - Get logs on nearly everything", inline=False)
+        embed.add_field(name="Admin", value="**leave** - Makes the bot leave the guild\n**leaveid** - Leaves a server by ID\n**setpresence(sp)** - Change the playing status of the bot.\n**shutdown** - Sends the bot into a deep sleep ...\n**cog** - Displays list of Cog Options\n**todo** - Displays List of shit todo\n**pm** - PMs Target user as bot\n**pmid** - PMs target ID as bot\n**amiadmin** - Tells you if your UserID is inside the cfg file.", inline=False)
         await author.send(embed=embed)
         await asyncio.sleep(10)
         await ctx.message.delete()
@@ -69,7 +74,7 @@ class Core(commands.Cog):
             )
         await user.send(embed=ee)
         eee = discord.Embed(
-            description = "__**Changelog (17/06/2019) v1.0.2**__\nAdded music command!(Play, Pause, Resume, volume, Stop)\n+ Added gif and gifr commands\n+ Added Hackban!\n+ Added pmid\n\n- Reworked the changelog command and put it in size order (iiCarelessness)\n- Reworked and updated Help command\n- Planted logos everywhere!\n\n__**Changelog (18/06/2019) v1.1**__\n+ Added a launcher gui with a few features\n+ Added Set Activity command\n+ Created a new admin cog\n+ Added amiadmin command\n+ Added utils folder\n+ Added config file\n\n- Merged lib into a new file named default inside util\n- Music now creates a folder for songs\n- Updated help command\n- Fixed some music bugs\n\n__**Changelog (18/06/2019) v1.1.1**__\n+ Added owo command (944)\n\n- Fixed Punish not setting channel permissions\n - Finished Cleanup command\n- Fixed volume command\n- Updated help command\n- Added clean command\n- Bug Fixes\n\n__**Changelog (21/06/2019) v1.2**__\n+ Added Error handler (catches and resolves errors automatically)\n+ Added help command for bot required permissions\n+ Added self delete function to every command\n+ Added 'role exist' check to remove and add\n+ Added sstop command (Force stop song)\n+ Added command cooldowns\n\n- Updated 'Forgot Something' errors to add more detail and to give a similar appearance\n- Reworked invite command (Invite ClientID is now based on the bots ID)\n- Changed stop command so only the song player can stop the song\n- Rewrote every command and optimized a lot of code\n- Rearranged and removed unused imports\n- Tweaked and tidied changelog output\n- Reverted and updated help command\n- Placed all commands into cogs\n- Reworked Cog loading system\n- Removed todo command\n- Reworked every cog\n- Reworked bot.py file\n- Bug Fixes",
+            description = "__**Changelog (17/06/2019) v1.0.2**__\nAdded music command!(Play, Pause, Resume, volume, Stop)\n+ Added gif and gifr commands\n+ Added Hackban!\n+ Added pmid\n\n- Reworked the changelog command and put it in size order (iiCarelessness)\n- Reworked and updated Help command\n- Planted logos everywhere!\n\n__**Changelog (18/06/2019) v1.1**__\n+ Added a launcher gui with a few features\n+ Added Set Activity command\n+ Created a new admin cog\n+ Added amiadmin command\n+ Added utils folder\n+ Added config file\n\n- Merged lib into a new file named default inside util\n- Music now creates a folder for songs\n- Updated help command\n- Fixed some music bugs\n\n__**Changelog (18/06/2019) v1.1.1**__\n+ Added owo command (944)\n\n- Fixed Punish not setting channel permissions\n - Finished Cleanup command\n- Fixed volume command\n- Updated help command\n- Added clean command\n- Bug Fixes\n\n__**Changelog (21/06/2019) v1.2**__\n+ Added Error handler (catches and resolves errors automatically)\n+ Added help command for bot required permissions\n+ Added self delete function to every command\n+ Added 'role exist' check to remove and add\n+ Added sstop command (Force stop song)\n+ Added command cooldowns\n\n- Updated 'Forgot Something' errors to add more detail and to give a similar appearance\n- Reworked invite command (Invite ClientID is now based on the bots ID)\n- Changed stop command so only the song player can stop the song\n- Rewrote every command and optimized a lot of code\n- Rearranged and removed unused imports\n- Tweaked and tidied changelog output\n- Reverted and updated help command\n- Placed all commands into cogs\n- Reworked Cog loading system\n- Removed todo command\n- Reworked every cog\n- Reworked bot.py file\n- Bug Fixes\n\n__**Changelog (21/06/2019) v1.3**__\n+Added customcommands\n+ Added leaveid\n+ Added logs\n\n- Added permission check to spp\n- Updated help command\n- Updated changelog\n- Bug fixes",
             colour = 0x9bf442,
             timestamp=datetime.datetime.utcnow()
             )
@@ -137,7 +142,7 @@ class Core(commands.Cog):
         embed.add_field(name="Server ID", value=guildid, inline=True)
         embed.add_field(name="Message", value=msg.content, inline=True)
         await me.send(embed=embed)
-        f = await ctx.send(embed=lib.Editable("Success!", f"Thanks to you another bug is about to be squished! Thank you for your feedback **{ctx.author.name}** :smile:", "Bug Report"))
+        f = await ctx.send(embed=lib.Editable("Success", f"Thanks to you another bug is about to be squished! Thank you for your feedback **{ctx.author.name}** :smile:", "Bug Report"))
         await lib.erase(ctx, 20, f)
         await ques.delete()
         await msg.delete()
@@ -145,24 +150,28 @@ class Core(commands.Cog):
     @commands.command(no_pm=True)
     @commands.cooldown(1, 60, commands.BucketType.guild)
     async def spp(self, ctx):
-        for channel in ctx.message.guild.channels:
-            role = discord.utils.get(channel.guild.roles, name="punished")
-            overwrite = discord.PermissionOverwrite()
-            overwrite.send_messages = False
-            overwrite.send_tts_messages = False
-            overwrite.add_reactions = False
-            await channel.set_permissions(role, overwrite=overwrite),
-        p = await ctx.send(embed=lib.Editable("Setting Permissions", "This may take a while, Ill tell you when im done.", "Moderation"))
-        await asyncio.sleep(5)
-        msg2 = await ctx.send(embed=lib.Editable("Im Finished!", "All permissions should be set.", "Moderation"))
-        await lib.erase(ctx, 5, p)
-        await asyncio.sleep(10)
-        await msg2.delete()
+        if ctx.author.guild_permissions.manage_channels:
+            for channel in ctx.message.guild.channels:
+                role = discord.utils.get(channel.guild.roles, name="punished")
+                overwrite = discord.PermissionOverwrite()
+                overwrite.send_messages = False
+                overwrite.send_tts_messages = False
+                overwrite.add_reactions = False
+                await channel.set_permissions(role, overwrite=overwrite),
+            p = await ctx.send(embed=lib.Editable("Setting Permissions", "This may take a while, Ill tell you when im done.", "Moderation"))
+            await asyncio.sleep(5)
+            msg2 = await ctx.send(embed=lib.Editable("Im Finished!", "All permissions should be set.", "Moderation"))
+            await lib.erase(ctx, 5, p)
+            await asyncio.sleep(10)
+            await msg2.delete()
+        else:
+            p = await ctx.send(embed=lib.NoPerm())
+            await lib.erase(ctx, 20, p)
 
     @commands.command(no_pm=True)
     @commands.cooldown(1, 5, commands.BucketType.user)
     async def music(self, ctx):
-        m = await ctx.send(embed=lib.Editable("Music Usage", "**play {song/link}** - Plays a song by name or url from youtube\n**pause** - Pauses the current song\n**resume** - Resumes the current song\n**volume {number}** - Change the volume of the bot\n**stop** - Disconnects the bot ", "Todo"))
+        m = await ctx.send(embed=lib.Editable("Music Usage", "**play {song/link}** - Plays a song by name or url from youtube\n**pause** - Pauses the current song\n**resume** - Resumes the current song\n**volume {number}** - Change the volume of the bot\n**stop** - Disconnects the bot\n**sstop** - Force disconnects the bot ", "Todo"))
         await lib.erase(ctx, 20, m)
 
     @commands.command(no_pm=True)
@@ -173,6 +182,160 @@ class Core(commands.Cog):
         await user.send(embed=lib.Editable("Github", "https://github.com/No1IrishStig/Devolution-Beta/", "Github"))
         await asyncio.sleep(10)
         await ctx.message.delete()
+
+    @commands.group(aliases=["cc"], invoke_without_command=True, no_pm=True)
+    @commands.cooldown(1, 5, commands.BucketType.user)
+    async def customcommand(self, ctx):
+        if ctx.author.guild_permissions.manage_messages:
+            await ctx.send(embed=lib.Editable("Custom Commands - Usage", "!cc add {name} {text}\n!cc edit {name} {text}\n!cc delete {name}\n!cc list\n\nAllows for the use of custom commands.", "Custom Commands"))
+        else:
+            p = await ctx.send(embed=lib.NoPerm())
+            await lib.erase(ctx, 20, p)
+
+    @customcommand.command(invoke_without_command=True, no_pm=True)
+    @commands.cooldown(1, 5, commands.BucketType.user)
+    async def add(self, ctx, command : str=None, *, text):
+        if ctx.author.guild_permissions.manage_messages:
+            server = ctx.guild
+            gid = str(server.id)
+            cc = self.cc
+            if command is None:
+                await ctx.send(embed=lib.Editable("Custom Commands - Usage", "!cc add {name} {text}\n\nCreate a custom command on this server.", "Custom Commands"))
+            else:
+                command = command.lower()
+                if command in self.bot.commands:
+                    await ctx.send(embed=lib.Editable("Error", "That command already exists!", "Error"))
+                    return
+                if gid not in cc:
+                    cc[gid] = {}
+                cmdlist = cc[gid]
+                if command not in cmdlist:
+                    cmdlist[command] = text
+                    cc[gid] = cmdlist
+                    with open("./data/customcommands/commands.json", "w") as f:
+                        json.dump(cc, f)
+                        await ctx.send(embed=lib.Editable("Success", "Custom command successfully added.", "Custom Commands"))
+                else:
+                    await ctx.send(embed=lib.Editable("Error", "This customcommand already exists. Use `!customcommand edit` to edit it.", "Error"))
+        else:
+            p = await ctx.send(embed=lib.NoPerm())
+            await lib.erase(ctx, 20, p)
+
+    @customcommand.command(invoke_without_command=True, no_pm=True)
+    @commands.cooldown(1, 5, commands.BucketType.user)
+    async def edit(self, ctx, command : str=None, *, text):
+        if ctx.author.guild_permissions.manage_messages:
+            server = ctx.guild
+            cc = self.cc
+            gid = str(server.id)
+            if command is None:
+                await ctx.send(embed=lib.Editable("Custom Commands - Usage", "!cc edit {name} {text}\n\nEdit a custom command on this server.", "Custom Commands"))
+            else:
+                command = command.lower()
+                if gid in cc:
+                    cmdlist = cc[gid]
+                    if command in cmdlist:
+                        cmdlist[command] = text
+                        cc[gid] = cmdlist
+                        with open("./data/customcommands/commands.json", "w") as f:
+                            json.dump(cc, f)
+                            await ctx.send(embed=lib.Editable("Success", "Custom command successfully edited.", "Custom Commands"))
+                    else:
+                        await ctx.send(embed=lib.Editable("Error", "That command doesn't exist. Use `!customcom add` to add it.", "Error"))
+                else:
+                    await ctx.send(embed=lib.Editable("Error", "There are no custom commands in this server. Use `!customcom add` to start adding some.", "Error"))
+        else:
+            p = await ctx.send(embed=lib.NoPerm())
+            await lib.erase(ctx, 20, p)
+
+    @customcommand.command(invoke_without_command=True, no_pm=True)
+    @commands.cooldown(1, 5, commands.BucketType.user)
+    async def delete(self, ctx, command:str=None):
+        if ctx.author.guild_permissions.manage_messages:
+            server = ctx.guild
+            cc = self.cc
+            gid = str(server.id)
+            if command is None:
+                await ctx.send(embed=lib.Editable("Custom Commands - Usage", "!cc delete {name}\n\nDelete a custom command on this server.", "Custom Commands"))
+            else:
+                command = command.lower()
+                if gid in cc:
+                    cmdlist = cc[gid]
+                    if command in cmdlist:
+                        cmdlist.pop(command, None)
+                        cc[gid] = cmdlist
+                        with open("./data/customcommands/commands.json", "w") as f:
+                            json.dump(cc, f)
+                            await ctx.send(embed=lib.Editable("Success", "Custom command successfully deleted.", "Custom Commands"))
+                    else:
+                        await ctx.send(embed=lib.Editable("Error", "That command doesn't exist.", "Error"))
+                else:
+                    await ctx.send(embed=lib.Editable("Error", "There are no custom commands in this server. Use `!customcom add` to start adding some.", "Error"))
+        else:
+            p = await ctx.send(embed=lib.NoPerm())
+            await lib.erase(ctx, 20, p)
+
+    @customcommand.command(invoke_without_command=True, no_pm=True)
+    @commands.cooldown(1, 5, commands.BucketType.user)
+    async def list(self, ctx):
+        if ctx.author.guild_permissions.manage_messages:
+            server = ctx.guild
+            cc = self.cc
+            gid = str(server.id)
+
+            commands = cc.get(gid, {})
+
+            if not commands:
+                await ctx.send(embed=lib.Editable("Error", "There are no custom commands for this server. Use `!customcommand add` to start adding some.", "Error"))
+                return
+
+            commands = ", ".join([ctx.prefix + c for c in sorted(commands)])
+            commands = "Custom commands:\n\n" + commands
+
+            if len(commands) < 1500:
+                await ctx.send(commands)
+            else:
+                for page in pagify(commands, delims=[" ", "\n"]):
+                    await ctx.author.send(page)
+        else:
+            p = await ctx.send(embed=lib.NoPerm())
+            await lib.erase(ctx, 20, p)
+
+    @commands.Cog.listener()
+    async def on_message(self, message):
+        try:
+            server = message.guild
+            gid = str(server.id)
+            cc = self.cc
+            prefix = "!"
+
+            if not prefix:
+                return
+
+            if gid in cc:
+                cmdlist = cc[gid]
+                cmd = message.content[len(prefix):]
+                if cmd in cmdlist:
+                    cmd = cmdlist[cmd]
+                    dest = message.channel
+                    await dest.send(cmd)
+                elif cmd.lower() in cmdlist:
+                    cmd = cmdlist[cmd.lower()]
+                    cmd = self.format_cc(cmd, message)
+                    dest = message.channel
+                    await dest.send(cmd)
+        except Exception as e:
+            return
+
+    def format_cc(self, command, message):
+        results = re.findall("\{([^}]+)\}", command)
+        for result in results:
+            param = self.transform_parameter(result, message)
+            command = command.replace("{" + result + "}", param)
+        return command
+
+
+
 
 def setup(bot):
     bot.add_cog(Core(bot))
